# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "iWell OpenApi Demo Documentation",
        "description": "Swagger OpenApi description",
        "contact": {
            "email": "yaroslav.bemov@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "article"
                ],
                "summary": "Get all articles",
                "description": "Get list of all articles or articles by user id if it is in query parameters",
                "operationId": "getArticles",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token",
                        "required": true
                    },
                    {
                        "name": "userId",
                        "in": "query",
                        "description": "User id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create new user",
                "description": "Create new user and return details",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "User first_name",
                        "required": true
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "User last_name",
                        "required": true
                    },
                    {
                        "name": "remember_token",
                        "in": "query",
                        "description": "User remember_token",
                        "required": false
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "User birthday",
                        "required": false
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "User gender",
                        "required": false
                    },
                    {
                        "name": "hometown_id",
                        "in": "query",
                        "description": "User hometown_id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/signin": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "signin",
                "description": "return user info and token",
                "operationId": "signin",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/signout": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "signout",
                "description": "revoke user token",
                "operationId": "signout",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "User token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/cities": {
            "get": {
                "tags": [
                    "city"
                ],
                "summary": "Get all cities",
                "description": "Get list of all cities",
                "operationId": "getCities",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/cities/{id}": {
            "get": {
                "tags": [
                    "city"
                ],
                "summary": "Get city by id",
                "description": "Get one city by id",
                "operationId": "getCity",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "City id",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/register/coach": {
            "post": {
                "tags": [
                    "coach"
                ],
                "summary": "Create new coach",
                "description": "Create new coach and return details",
                "operationId": "registerCoach",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Coach email",
                        "required": true
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Coach password",
                        "required": true
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "Coach first_name",
                        "required": true
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Coach last_name",
                        "required": true
                    },
                    {
                        "name": "remember_token",
                        "in": "query",
                        "description": "Coach remember_token",
                        "required": false
                    },
                    {
                        "name": "birthday",
                        "in": "query",
                        "description": "Coach birthday",
                        "required": false
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Coach gender",
                        "required": false
                    },
                    {
                        "name": "hometown_id",
                        "in": "query",
                        "description": "Coach hometown_id",
                        "required": false
                    },
                    {
                        "name": "achieve",
                        "in": "query",
                        "description": "Coach achieve",
                        "required": false
                    },
                    {
                        "name": "rank",
                        "in": "query",
                        "description": "Coach rank",
                        "required": false
                    },
                    {
                        "name": "about",
                        "in": "query",
                        "description": "Coach about",
                        "required": false
                    },
                    {
                        "name": "skill_id",
                        "in": "query",
                        "description": "Coach skill_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/skills": {
            "get": {
                "tags": [
                    "coach"
                ],
                "summary": "Get all skills",
                "description": "Get list of all skills",
                "operationId": "getSkills",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/skill-types": {
            "get": {
                "tags": [
                    "coach"
                ],
                "summary": "Get all Skill Types",
                "description": "Get list of all Skill Types",
                "operationId": "getSkillTypes",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Access token",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "iWell",
            "description": "API Endpoints of iWell"
        }
    ],
    "security": [
        []
    ]
}